[IMP] core: move json_default from tools/date_utils to tools/json
Since odoo/odoo@1ecb0641efead9966faafb853c1e03138a49f111
`json_default` is no longer entirely related to dates anymore.

Besides, if you want to a new object to implement the JSON
serialization for, which has nothing to do with dates,
it's weird to have to add it in tools/date_utils.

I therefore suggest to move this in tools/json rather than tools/date_utils
