+++ odoo/addons/test_performance/tests/test_performance.py
@@ -7,7 +7,7 @@ import logging
 
 from odoo.addons.base.tests.common import SavepointCaseWithUserDemo
 from odoo.tests.common import TransactionCase, users, warmup, tagged
-from odoo.tools import mute_logger, json_default, sql
+from odoo.tools import mute_logger, sql
 from odoo import Command
 
 _logger = logging.getLogger(__name__)
